+++ src/main/java/com/github/jlangch/venice/impl/docgen/cheatsheet/DocGenerator.java
@@ -44,6 +44,7 @@ import com.github.jlangch.venice.impl.docgen.cheatsheet.modules.ModuleBenchmarkS
 import com.github.jlangch.venice.impl.docgen.cheatsheet.modules.ModuleComponentSection;
 import com.github.jlangch.venice.impl.docgen.cheatsheet.modules.ModuleConfigSection;
 import com.github.jlangch.venice.impl.docgen.cheatsheet.modules.ModuleCryptographySection;
+import com.github.jlangch.venice.impl.docgen.cheatsheet.modules.ModuleDockerSection;
 import com.github.jlangch.venice.impl.docgen.cheatsheet.modules.ModuleExcelSection;
 import com.github.jlangch.venice.impl.docgen.cheatsheet.modules.ModuleFontsSection;
 import com.github.jlangch.venice.impl.docgen.cheatsheet.modules.ModuleGeoipSection;
@@ -118,7 +119,7 @@ public class DocGenerator {
                         "java",   "semver",   "excel",     "hexdump",
                         "shell",  "geoip",    "benchmark", "component",
                         "config", "parsifal", "grep",      "test",
-                        "fonts",  "qrref"));
+                        "fonts",  "qrref",    "docker"));
 
         final IVeniceInterpreter venice = new VeniceInterpreter(new AcceptAllInterceptor());
 
@@ -310,6 +311,7 @@ public class DocGenerator {
         extmod.addSection(new DocSection("Ansi", "modules.ansi"));
         extmod.addSection(new DocSection("Gradle", "modules.gradle"));
         extmod.addSection(new DocSection("Maven", "modules.maven"));
+        extmod.addSection(new DocSection("Docker", "modules.docker"));
         extmod.addSection(new DocSection("Test", "modules.test"));
         extmod.addSection(new DocSection("Tracing", "modules.tracing"));
         extmod.addSection(new DocSection("Benchmark", "modules.benchmark"));
@@ -390,6 +392,7 @@ public class DocGenerator {
                 new ModuleParsifalSection(diBuilder).section(),
                 new ModuleGradleSection(diBuilder).section(),
                 new ModuleMavenSection(diBuilder).section(),
+                new ModuleDockerSection(diBuilder).section(),
                 new ModuleTracingSection(diBuilder).section(),
                 new ModuleShellSection(diBuilder).section(),
                 new ModuleAnsiSection(diBuilder).section(),