+++ src/main/java/com/github/jlangch/venice/impl/modules/Modules.java
@@ -56,6 +56,7 @@ public class Modules {
                         "component",
                         "crypt",
                         "dag",
+                        "docker",
                         "docx",
                         "excel",
                         "excel-install",