+++ src/main/java/com/github/jlangch/venice/impl/namespaces/Namespaces.java
@@ -114,7 +114,7 @@ public class Namespaces {
                     Arrays.asList(
                         "bench",
                         "component", "crypt",
-                        "docx",
+                        "docker", "docx",
                         "excel",
                         "http",
                         "jackson",